refactor : Integer 타입을 Long 타입으로 변경

+++ src/main/java/com/bootcamp/paymentdemo/domain/order/dto/OrderCreateRequest.java
@@ -9,11 +9,11 @@ import java.util.List;
 public class OrderCreateRequest {
 
     private Long memberId;
-    private Integer totalAmount;
-    private Integer usedPoints;
-    private Integer finalAmount;
-    private Integer earnedPoints;
-    private Integer quantity;
+    private Long totalAmount;
+    private Long usedPoints;
+    private Long finalAmount;
+    private Long earnedPoints;
+    private Long quantity;
     private String currency;
 
     private List<OrderItemRequest> items = new ArrayList<>();
